Cleanup to find bug (#39)
* Cleanup to find bug

Standardize flags, parse IPs in a more cautious manner, don't run
local-local traceroutes, add a dryRun flag.

* Improved coverage for connectionpoller

* Remove coverpkg stuff

* Fix coverage

* Made dry-run functionality work.

+++ scamper/scamper.go
@@ -29,6 +29,8 @@ import (
 // and managed.
 type Daemon struct {
 	Binary, AttachBinary, Warts2JSONBinary, ControlSocket, OutputPath string
+
+	DryRun bool
 }
 
 var (
@@ -201,14 +203,19 @@ func (d *Daemon) trace(conn connection.Connection, t time.Time) string {
 	_, err := buff.WriteString(GetMetaline(conn, 0, ""))
 	rtx.PanicOnError(err, "Could not write to buffer")
 
-	cmd := pipe.Line(
-		pipe.Println("tracelb -P icmp-echo -q 3 -O ptr ", conn.RemoteIP),
-		pipe.Exec(d.AttachBinary, "-i-", "-o-", "-U", d.ControlSocket),
-		pipe.Exec(d.Warts2JSONBinary),
-		pipe.Write(&buff),
-	)
-	rtx.PanicOnError(pipe.Run(cmd), "Command %v failed", cmd)
-	rtx.PanicOnError(ioutil.WriteFile(filename, buff.Bytes(), 0666), "Could not save output to file")
+	log.Printf(
+		"Running: echo \"tracelb -P icmp-echo -q 3 -O ptr %s\" | %s -i- -o- -U %s | %s > %s\n",
+		conn.RemoteIP, d.AttachBinary, d.ControlSocket, d.Warts2JSONBinary, filename)
+	if !d.DryRun {
+		cmd := pipe.Line(
+			pipe.Println("tracelb -P icmp-echo -q 3 -O ptr ", conn.RemoteIP),
+			pipe.Exec(d.AttachBinary, "-i-", "-o-", "-U", d.ControlSocket),
+			pipe.Exec(d.Warts2JSONBinary),
+			pipe.Write(&buff),
+		)
+		rtx.PanicOnError(pipe.Run(cmd), "Command %v failed", cmd)
+		rtx.PanicOnError(ioutil.WriteFile(filename, buff.Bytes(), 0666), "Could not save output to file")
+	}
 	return string(buff.Bytes())
 }
